requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-admin-export-action==0.3.1
django-adminlte-3==0.1.6
django-ckeditor==6.7.0
django-classy-tags==4.1.0
django-cors-headers==4.2.0
django-csp==3.7
django-grappelli==3.0.7
django-import-export==3.3.5
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-sekizai==4.1.0
django-taggit==4.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
